pass getDiscoverHref callback for consistency with other action props
